Moving forward with hydrating from multiple recordsets.
Added Type property to IMemberInfo and the classes that implement it
Some (non-working) code to run the next recordset.

This is recursive, so it will work for multiple levels, but the
stored procedures are going to get more complicated as to the correct
ordering of the recordsets to be retrieved.  Might need to make this into
a tail-recursion situation, or make it work only one level deep.